Use data keys on macOS for RSA, EC cryptography
